[MINOR] Introduce `miniagents/promising/promise_utils.py`; read `FROZEN_LONGER_HASH_KEYS` env var (#172)
* uv lock --upgrade

* stop serializing sub-message hashes (for now)

* minor

* allow promises inside Frozen; always store nested messages as MessagePromise

* move `cached_privately` to `promise_utils.py`; attempt to add custom `pydantic` serialization to `Promise`

* revert the first Promise in Frozen attempt

* in progress: support Promise inside Frozen

* prepare to serialize Promise inside Frozen

* introduce FROZEN_LONGER_HASH_KEYS env var

* type hints for _pydantic_serialize

* revert my attempt to implement Promise inside Message for now; uv lock --upgrade

* revert my attempt to implement Promise inside Message for now
