+++ miniagents/promising/promise_utils.py
@@ -1,7 +1,37 @@
 import asyncio
+from functools import wraps
 import inspect
 from asyncio import Future
-from typing import Any, Optional, Union
+from typing import Any, Callable, Optional, Union
+
+from miniagents.promising.sentinels import NO_VALUE
+
+
+def cached_privately(func: Callable[[Any], Any]) -> Callable[[Any], Any]:
+    """
+    Unlike `@functools.cached_property`, this decorator caches the result of the method call in a private attribute
+    instead of replacing the original method with the calculated value. This approach prevents the cached value from
+    being registered as a field value in the Pydantic model upon evaluation.
+
+    NOTE: This decorator does not automatically turn the method into a property - you need to additionally decorate
+    your method with `@property` on top of this decorator. This decision was made because IDEs like PyCharm don't seem
+    to realize that the method became a property if it wasn't explicitly decorated with known decorators like
+    `@property` or `@functools.cached_property` (they might have hardcoded this behaviour).
+    """
+
+    # TODO can it be made thread-safe ? (for the sake of tricks like `asyncio.to_thread()` and similar)
+
+    @wraps(func)
+    def wrapper(self: Any) -> Any:
+        # TODO [MINOR] Avoid dynamic construction of the field name upon each call ?
+        attr_name = f"__{type(self).__name__}__{func.__name__}__cache"
+        result = getattr(self, attr_name, NO_VALUE)
+        if result is NO_VALUE:
+            result = func(self)
+            setattr(self, attr_name, result)
+        return result
+
+    return wrapper
 
 
 async def acancel_async_object(
